rename clearingApp.ts to marginApp.ts
